packages/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "xujingzhou", email = "max3.xu@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "core",
    "confz>=2.1.0",
    "loguru>=0.7.3",
    "sqlalchemy>=2.0.44",
    "psycopg[binary]>=3.2.12",
    "fastapi[standard]>=0.120.4",
    "uvicorn>=0.38.0",
    "python-jose>=3.5.0",
    "ag-ui-protocol>=0.1.10",
    "agentscope>=1.0.9",
]

[tool.uv.sources]
core = { workspace = true }

[project.scripts]
api = "api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
